pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "google-cloud-texttospeech>=2.31.0",
    "google-genai>=1.41.0",
    "sift-stack-py>=0.9.1",
]
